+++ README.md
@@ -1,0 +1,9 @@
+# AlphaHReview Methodology
+
+### Systematic Literature Review Methodology
+
+This methodology was developed with the aim of prioritizing publication selection criteria and exposing
+alternatives to improve the systematic literature review method. The AHP tool is used as a methodology
+to classify the quality criteria of the most relevant publications for the researched topic.
+Furthermore, the Weitendorf ratios was used to consider and classify the most relevant publications for
+the researched topic.
