+++ css/styles.css
@@ -1,0 +1,261 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap');
+
+:root{
+    font-size: 62.5%;
+}
+
+/* Estilos Gerais */
+body {
+    font-family: Roboto, Arial, sans-serif;
+    font-size: 1.6rem;
+    line-height: 1.4;
+    margin: 0 auto;
+    padding: 5px;
+    background-color: #d3e6d3;
+    max-width: 1440px;
+}
+
+#title {
+    color: #42ac34;
+    text-align: center;
+}
+
+h1, h2, h3 {
+    color: #333;
+    text-align: center;
+}
+
+.campo {
+    margin-top: 2px;
+    margin-bottom: 2px;
+    padding: 5px;
+    background-color: #fff;
+    border: 1px solid #b9b9b9;
+    border-radius: 5px;
+}
+
+input, .uploadFile {
+    border: 1px solid #1b6828;
+    padding: 6px;
+    border-radius: 4px;
+}
+
+.input-criteria {
+    width: 100px;
+    margin-right: 10px;
+    margin-bottom: 10px;
+}
+
+.button, .stepbutton, .download-btn, select#criteria-number {
+    padding: 10px 20px;
+    background-color: #1b6828;
+    color: #fff;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    font-weight: bold;
+    transition: background-color 0.3s ease;
+    text-align: center;
+}
+
+.button:hover, .stepbutton:hover, .download-btn:hover, button:hover {
+    background-color: #278236;
+}
+
+.stepbutton {
+    display: block;
+    margin: 5px auto;
+}
+
+.container {
+    padding: 10px;
+    background-color: #f0f0f0;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    display: flex;
+    justify-content: center;
+}
+
+table, .table-comparison, #ahpTable {
+    border-collapse: collapse;
+    width: 100%;
+    margin: 20px 0;
+    background-color: #f9f9f9;
+    font-size: 1.2rem;
+    font-weight: bold;
+}
+
+#resultsContainer, .comparison-matrix {
+    margin-top: 20px;
+    margin-bottom: 20px;
+}
+
+#criterias, p, span, .small-text, .criteria {
+    font-size: 1.6rem;
+    margin: 5px 0;
+}
+
+th, td {
+    padding: 8px;
+    text-align: center;
+    border-bottom: 1px solid #ddd;
+}
+
+th {
+    font-size: 1.4rem;
+    background-color: #1b6828;
+    color: white;
+}
+
+td.criteria-label {
+    font-weight: bold;
+}
+
+.table-comparison .centered-cell {
+    text-align: center;
+}
+
+th.criteria-label, .criteria-label {
+    background-color: #f0f0f0;
+    font-weight: bold;
+    text-align: center;
+}
+
+tr:first-child {
+    background-color: #1b6828;
+    font-weight: bold;
+}
+
+tr:nth-child(even) {
+    background-color: #f9f9f9;
+}
+
+tr:nth-child(odd) {
+    background-color: #ffffff;
+}
+
+.highlighted-cell {
+    background-color: #278236;
+    color: #fff;
+}
+
+.border-bottom-row {
+    border-bottom: 3px solid #000;
+    color: #1b6828;
+}
+
+.consistent {
+    color: green;
+}
+
+.inconsistent {
+    color: red;
+}
+
+.textarea-custom {
+    padding: 10px;
+    border: 2px solid #1b6828;
+    border-radius: 5px;
+    font-size: 1.4rem;
+    width: 100%;
+    box-sizing: border-box;
+}
+
+#ahpTable {
+    max-width: 600px;
+    width: 90%;
+    margin: 0 auto;
+    margin-top: 10px;
+    margin-bottom: 10px;
+}
+
+#ahpTable th:first-child, #ahpTable td:first-child {
+    border-left: 1px solid #1b6828;
+}
+
+#ahpTable th, #ahpTable td {
+    border-bottom: 1px solid #1b6828;
+    border-right: 1px solid #1b6828;
+    border-left: 1px solid #1b6828;
+}
+
+#ahpTable tr:first-child th {
+    border-left: 1px solid #000000;
+    border-right: 1px solid #000000;
+    border-top: 1px solid #000000;
+    border-bottom: 2px solid #000000;
+}
+
+.caption {
+    font-size: 1.0rem;
+    font-weight: bold;
+    margin-bottom: 10px;
+}
+
+.small-text {
+    font-size: 1.3rem;
+    align-content: center;
+    padding-top: 5px;
+}
+
+.criteria-label {
+    width: 70px;
+}
+
+#selectedCriteria {
+    font-weight: bold;
+    margin-top: 10px;
+}
+
+.content {
+    text-align: center;
+}
+
+hr {
+    border: none;
+    border-top: 1px solid #ccc;
+    margin: 20px 0;
+}
+
+.comparison-matrix, .comparison-row {
+    display: flex;
+    flex-direction: column;
+    margin-top: 20px;
+}
+
+.comparison-row {
+    display: flex;
+    align-items: center;
+    margin-bottom: 10px;
+}
+
+#resultsTable td:nth-child(4) {
+    border-right: 2px solid #adadad;
+}
+
+#resultsTable tr:first-child, #resultsTable tr:first-child {
+    background-color: #1b6828;
+    color: white;
+}
+
+#resultsTable tr:nth-last-child(3) td {
+    border-bottom: 1px solid #1b6828;
+}
+
+#resultsTable  tr:nth-last-child(1) td {
+    border-bottom: 2px solid #1b6828;
+}
+
+
+#resultsWT td:nth-child(4) {
+    border-right: 2px solid #1b6828;
+}
+
+#resultsWT tr:first-child, #resultsWT tr:first-child {
+    background-color: #1b6828;
+    color: white;
+}
+
+#resultsWT tr:nth-last-child(1) td {
+    border-bottom: 2px solid #1b6828;
+}
